+++ ChangeLog.md
@@ -1,6 +1,7 @@
 #2.x
   * Fix material name for Cadmium
-  * Add ability to set state variables from VoiceAttack
+  * Add ability to set state variables from VoiceAttack.  Details on how to use this are in the VoiceAttack documentation
+  * Add option to write speech responder output to a file.  This is an option that can be checked in the speech responder tab, and writes all speech to %APPDATA%\EDDI\speechresponder.out
 
 #2.0.13
   * Fix issue where engineer rank journal message without rank would cause a crash